Convert HTML to Markdown
The site is now generated from Markdown source files instead of raw
HTML. To generate the site you need to install Jekyll, please see:

  https://help.github.com/articles/using-jekyll-with-pages

When Jekyll is installed, you generate the site with

  % jekyll serve --watch

This will start a webserver on http://localhost:4000/ where you can
see the site. Changes in the Markdown files are immediatedly visible
in the test server.