Use SCSS for main CSS file
SCSS is a CSS pre-processor that allows for things like defining
variables. Jekyll automatically converts SCSS files into CSS files.

This is a simple rename, further changes will clean up the file and
begin using SCSS features.

The two "---" lines at the top are the trigger for Jekyll to process
the file. It is an empty YAML front-matter.