Refactor main SCSS file
This is a simple "run-length encoding" transformation of the SCSS file
where adjacent and repeated rules were turned into nested rules. So

  foo bar { ... }
  foo baz { ... }

became

  foo {
      bar { ... }
      baz { ... }
  }

The integrity of this change was checked by comparing the output of
cssunminify main.css before and after. Later refactorings will reorder
the rules to clean things up some more.